Allowed user to leave preferences blank for default values

+++ app/src/main/java/ca/dal/csci3130/quickcash/usermanagement/PreferenceInterface.java
@@ -14,8 +14,8 @@ public interface PreferenceInterface {
 
     void setPayRate(double payRate);
 
-    int getDuration();
+    double getDuration();
 
-    void setDuration(int duration);
+    void setDuration(double duration);
 
 }
